feat: pagination via offset in server route
