+++ package.json
@@ -63,11 +63,10 @@
     ]
   },
   "engines": {
-    "node": "16.16.0"
+    "node": "16.x"
   },
   "devDependencies": {
     "@types/lodash": "^4.14.182",
-    "@types/react-copy-to-clipboard": "^5.0.4",
-    "eslint-plugin-react": "^7.30.1"
+    "@types/react-copy-to-clipboard": "^5.0.4"
   }
 }
