+++ server/queries.ts
@@ -16,7 +16,7 @@ import { ILine, IPoem } from "../src/types";
     }
 })();
 
-async function returnPoems(nPoems: number) {
+async function returnPoems(nPoems: number, offset = 0) {
     try {
         await Poem.sync();  // make sure the table exists
         await Line.sync();  // make sure the table exists
@@ -28,6 +28,7 @@ async function returnPoems(nPoems: number) {
                 "content",
             ],
             limit: nPoems,
+            offset,
             order: [
                 [
                     "createdAt",
