+++ src/helpers/index.ts
@@ -1,5 +1,4 @@
 import * as React from "react";
-import { IPoem } from "../types";
 
 const allPossibleLetters = "abcdefghijklmnopqrstuvwxyz";
 const quantityAllPossibleLetters = allPossibleLetters.length;
@@ -25,20 +24,6 @@ export function alphaCharacterRotate(text: string) {
     }
 
     return a.join("");
-}
-
-export function getItemsOnCurrentPage(items: IPoem[], page: number, itemsPerPage: number): IPoem[] {
-    const { length } = items;
-    const firstIndex = (page - 1) * itemsPerPage;
-    const lastIndex = length < itemsPerPage
-        ? length
-        : firstIndex + itemsPerPage;
-    
-    const itemsOnCurrentPage: IPoem[] = [];
-    for (let i = firstIndex; i < lastIndex; i++) {
-        itemsOnCurrentPage.push(items[i]);
-    }
-    return itemsOnCurrentPage;
 }
 
 // this function allows us to get the most current value of a state variable
